move test_create_brick_valid into separate file test_brick.py
